requirements.txt
Authlib==1.2.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
dominate==2.7.0
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==3.0.2
Flask==2.2.2
greenlet==2.0.1
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
protobuf==4.21.12
pycparser==2.21
SQLAlchemy==1.4.45
Tkinter==0.0.0
visitor==0.1.3
Werkzeug==2.2.2
zipp==3.11.0
